[PRMDR-577] Add stage for feedback form to ease latter ticket
