+++ app/src/components/blocks/feedbackConfirmation/FeedbackConfirmation.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function FeedbackConfirmation() {
+    return (
+        <>
+            <h1>We’ve received your feedback</h1>
+            <p>If you have left your details, our team will contact you soon. </p>
+            <p>You can now close this window.</p>
+        </>
+    );
+}
+
+export default FeedbackConfirmation;
