+++ app/src/helpers/requests/sendEmail.ts
@@ -4,6 +4,7 @@ const sendEmail = async (formData: FormData) => {
     // using console.log as a placeholder until we got the send email solution in place
     /* eslint-disable-next-line no-console */
     console.log(`sending feedback from user by email: ${JSON.stringify(formData)}}`);
+    await new Promise((resolve) => setTimeout(resolve, 1000));
     return { status: 200 };
 };
